ml-worker/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.32.0
numpy==2.1.0
onnxruntime==1.20.0
sounddevice==0.5.0
soundfile==0.12.1
faster-whisper==1.1.0
silero-vad==5.1.2
pyannote.audio==3.3.2
resemblyzer==0.1.4
sentence-transformers==3.3.1
psutil==6.1.0
httpx==0.28.1
sse-starlette==2.2.1
